requirements.txt
PyGithub==1.56
requests==2.28.1
